Move tests to __tests__; Prov to ProvData
