+++ context/app/static/js/components/ProvVis/ProvVis.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 
 import Graph, { GraphParser } from '@hms-dbmi-bgm/react-workflow-viz';
 
-import Prov from './Prov';
+import ProvData from './ProvData';
 
 export default function ProvVis(props) {
   const { prov, getNameForActivity, getNameForEntity, renderDetailPane } = props;
-  const steps = new Prov(prov, getNameForActivity, getNameForEntity).toCwl();
+  const steps = new ProvData(prov, getNameForActivity, getNameForEntity).toCwl();
   // eslint-disable-next-line consistent-return
   function renderDetailPaneWithNode(node) {
     if (renderDetailPane && node) {
