+++ context/app/static/js/components/ProvVis/__tests__/fixtures/complex-cwl.json
