+++ context/app/static/js/components/ProvVis/__tests__/fixtures/primary.cwlprov.json
