+++ context/app/static/js/components/ProvVis/__tests__/fixtures/real-cwl.json
