+++ context/app/static/js/components/ProvVis/__tests__/fixtures/simple-cwl.json
